重写注释，补充README

+++ include/.general.h
@@ -8,11 +8,12 @@
 
 // 错误返回值
 #define ERROR 0x1FFF
-// 下标
+
+// 下标类型 (由于可能为-1, 因此不能为unsigned类型)
 typedef long ind_t;
-// 长度
-typedef long len_t;
+// 长度类型
+typedef size_t len_t;
 // 全局数据类型
-typedef long ElemType;
+typedef int ElemType;
 
 #endif
